Change storage of physical address of shared mem
When a client of the TEE driver maps a shared buffer, the physical
address is stored in the kernel structure 'vm_area_struct'. The field
used is vm_pgoff. This field is updated by the kernel on the
remap_pfn_ranhe function. However in the kernel 3.10, this update is
done only if the piece of memory is copy-on-write mapped. In our use
case, we do not support the copy-on-write, so we can't use the vm_pgoff
field. So the physical address is now stored in the vm_private_data
field.

Change-Id: Ie67af3bdc0c21dcc21ae7512dda4d9c6cd332813
Signed-off-by: Emmanuel Michel <emmanuel.michel@st.com>
Reviewed-on: https://gerrit.st.com/6540
Reviewed-by: Jean-Michel DELORME <jean-michel.delorme@st.com>