Support 32-bit client working with 64-bit linux kernel
Since compat ioctl is not supported by current driver, 32-bit client
application will fail to call ioctl if linux kernel is 64-bit.

Meanwhile, client and driver are sharing lots of data struct which is
unaligned in 64-bit kernel and those struct is widely used by *sizeof*
operator both in client and driver.
If client is 32-bit, the data size may be totally different in 64-bit
kernel and data copy between client and driver will be failed.

This commit includes
- add compat ioctl
- restructure unaligned struct and add padding member

Signed-off-by: Aijun Sun <aijun.sun@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey)
Tested-by: Aijun Sun <aijun.sun@linaro.org> (FVP platform: 32/64-bit xtest)
Tested-by: Aijun Sun <aijun.sun@linaro.org> (Qemu platform)