Allow the driver to be linked into the kernel image
With the minimal Kbuild infrastructure to bring the OPTEE
linuxdriver in kernel tree, the following build error occurs
when configuring the driver as builtin into the kernel:

<path-to>/armtz/built-in.o:(.rodata+0x50):
  multiple definition of `tee_fops'
<path-to>/core/built-in.o:(.rodata+0x60):
  first defined here

This is due to name clash and is not reproduced when OPTEE is built as
loadable modules. Fix it by declaring the fops structures as static.

Signed-off-by: Eugeniu Rosca <erosca@de.adit-jv.com>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>